refac(Tour): extract TourHelper to its own file
